Rename TextBox for consistency, remove now-unnecessary character module.
